feat(project): add project card grid and list views, enhance project stats and quick actions -> fix: dashboard errors
